Move LocalDynamoDB into the new storage package
